+++ libs/PCLEngine/VTKExtensions/Widgets/ScaleBar.h
@@ -18,7 +18,9 @@
 
 // Qt version compatibility handling
 #include <QApplication>
+#if QT_VERSION < QT_VERSION_CHECK(6, 0, 0)
 #include <QDesktopWidget>
+#endif
 #include <QScreen>
 
 class ScaleBar {